refactor(cli): split Sigma ACP agent responsibilities
